Reformat all code
