+++ cyfhir/src/main/java/com/Optum/CyFHIR/CyFhir.java
@@ -1,0 +1,7 @@
+package com.Optum.CyFHIR;
+
+public class CyFhir {
+
+    public static void main(String[] args) {}
+
+}
