+++ cyfhir/src/main/java/com/Optum/CyFHIR/models/FhirRecursiveObj.java
@@ -1,0 +1,27 @@
+package com.Optum.CyFHIR.models;
+
+
+import org.neo4j.graphdb.Node;
+
+import java.util.ArrayList;
+
+public class FhirRecursiveObj {
+    Node node;
+    ArrayList<FhirRelationship> relationships;
+
+    public Node getNode() {
+        return node;
+    }
+
+    public ArrayList<FhirRelationship> getRelationships() {
+        return relationships;
+    }
+
+    public void setNode(Node node) {
+        this.node = node;
+    }
+
+    public void setRelationships(ArrayList<FhirRelationship> relationships) {
+        this.relationships = relationships;
+    }
+}
